+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning].
 - Add JS helpers.
 
 ```js
-import { cableUrl, turboStreamName } from 'http://anycable.io/xk6-cable/jslib/k6-rails/0.1.0/index.js'
+import { cableUrl, turboStreamSource } from 'https://anycable.io/xk6-cable/jslib/k6-rails/0.1.0/index.js'
 
 export default function () {
   let res = http.get("http://localhost:3000/home");
@@ -19,9 +19,9 @@ export default function () {
   const wsUrl = cableUrl(html);
   let client = cable.connect(wsUrl);
 
-  let streamName = turboStreamName(html);
+  let { streamName, channelName } = turboStreamSource(html);
 
-  let channel = client.subscribe("Turbo::StreamsChannel", {
+  let channel = client.subscribe(channelName, {
     signed_stream_name: streamName,
   });